Move emojify into filter subpackage
