+++ app/shell/py/pie/tests/test_emojify.py
@@ -1,6 +1,6 @@
 import sys
 
-from pie import emojify
+from pie.filter import emojify
 
 
 def test_emojify_replaces_aliases(capsys):
